feat: extract version creation to separate file per pkg
