+++ packages/wdio-cli/.gitignore
@@ -1,0 +1,1 @@
+src/version.ts
