+++ packages/wdio-cli/src/run.ts
@@ -6,7 +6,8 @@ import { hideBin } from 'yargs/helpers'
 
 import { commands } from './commands/index.js'
 import { handler, cmdArgs } from './commands/run.js'
-import { CLI_EPILOGUE, packageVersion } from './constants.js'
+import { CLI_EPILOGUE } from './constants.js'
+import { PKG_VERSION } from './version.js'
 import type { RunCommandArguments } from './types.js'
 
 const DEFAULT_CONFIG_FILENAME = 'wdio.conf.js'
@@ -34,7 +35,7 @@ export default async function run() {
         .example('wdio install reporter spec', 'Install @wdio/spec-reporter')
         .example('wdio repl chrome -u <SAUCE_USERNAME> -k <SAUCE_ACCESS_KEY>', 'Run repl in Sauce Labs cloud')
         .updateStrings({ 'Commands:': `${DESCRIPTION.join('\n')}\n\nCommands:` })
-        .version(packageVersion)
+        .version(PKG_VERSION)
         .epilogue(CLI_EPILOGUE)
 
     /**
