+++ tests/helpers/launch.js
@@ -1,4 +1,4 @@
-import Launcher from '../../packages/wdio-cli/build/src/launcher.js'
+import Launcher from '../../packages/wdio-cli/build/launcher.js'
 
 export default function launch (testName, ...args) {
     const launcher = new Launcher(...args)
